Moved all files in right packages and added spring boot framework
