--- api/src/main/kotlin/ApiObjectMapper.kt
@@ -1,6 +1,0 @@
-import com.fasterxml.jackson.databind.DeserializationFeature
-import com.fasterxml.jackson.databind.ObjectMapper
-
-val apiObjectMapper = ObjectMapper().apply {
-    disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
-}
