+++ api/src/main/kotlin/com/crypto/api/ApiObjectMapper.kt
@@ -1,0 +1,8 @@
+package com.crypto.api
+
+import com.fasterxml.jackson.databind.DeserializationFeature
+import com.fasterxml.jackson.databind.ObjectMapper
+
+val apiObjectMapper = ObjectMapper().apply {
+    disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
+}
