+++ build.gradle.kts
@@ -2,8 +2,10 @@ import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
     kotlin("jvm")
-    id("io.gitlab.arturbosch.detekt").version("1.22.0")
+    id("io.gitlab.arturbosch.detekt").version("1.22.0") // TODO ?
 }
+
+val JVM_TARGET = "17"
 
 group = "com.crypto"
 version = "0.0.1"
@@ -19,21 +21,24 @@ subprojects {
     version = rootProject.version
 
     tasks.withType<KotlinCompile> {
-        kotlinOptions.jvmTarget = "17"
+        kotlinOptions.jvmTarget = JVM_TARGET
+    }
+    tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinJvmCompile> {
+        kotlinOptions.jvmTarget = JVM_TARGET
     }
 }
 
-//detekt {
-//    source = files("common")
-//    buildUponDefaultConfig = true // preconfigure defaults
-//    allRules = false // activate all available (even unstable) rules.
-//    config = files("$projectDir/detekt.yml") // point to your custom config defining rules to run, overwriting default behavior
-////    baseline = file("$projectDir/baseline.xml") // a way of suppressing issues before introducing detekt
-//}
+detekt {
+    source = files("common", "api", "mappers", "order-app-spring")
+    buildUponDefaultConfig = true // preconfigure defaults
+    allRules = false // activate all available (even unstable) rules.
+    config = files("$projectDir/detekt.yml") // point to your custom config defining rules to run, overwriting default behavior
+    baseline = file("$projectDir/baseline.xml") // a way of suppressing issues before introducing detekt
+}
 
 tasks.withType<io.gitlab.arturbosch.detekt.Detekt>().configureEach {
-    jvmTarget = "17"
+    jvmTarget = JVM_TARGET
 }
 tasks.withType<io.gitlab.arturbosch.detekt.DetektCreateBaselineTask>().configureEach {
-    jvmTarget = "17"
+    jvmTarget = JVM_TARGET
 }
