--- common/src/main/kotlin/CommonConstants.kt
@@ -1,3 +1,0 @@
-//private val INSTANT_NONE = Instant.fromEpochMilliseconds(Long.MIN_VALUE) // TODO it doesn't work
-//val Instant.Companion.NONE
-//    get() = INSTANT_NONE
