+++ common/src/main/kotlin/com/crypto/common/models/CommonCommand.kt
@@ -1,0 +1,9 @@
+package com.crypto.common.models
+
+enum class CommonCommand {
+    NONE,
+    CREATE,
+    READ,
+    DELETE,
+    SEARCH
+}
