+++ common/src/main/kotlin/com/crypto/common/models/CommonError.kt
@@ -1,0 +1,9 @@
+package com.crypto.common.models
+
+class CommonError(
+    val code: String = "",
+    val group: String = "",
+    val field: String = "",
+    val message: String = "",
+    val exception: Throwable? = null
+)
