+++ common/src/main/kotlin/com/crypto/common/models/CommonFilterOrder.kt
@@ -1,0 +1,7 @@
+package com.crypto.common.models
+
+data class CommonFilterOrder(
+    var type: CommonOrderType = CommonOrderType.NONE,
+    var operation: CommonOrderOperation = CommonOrderOperation.NONE,
+    var status: CommonOrderStatus = CommonOrderStatus.NONE,
+)
