+++ common/src/main/kotlin/com/crypto/common/models/CommonOrderId.kt
@@ -1,0 +1,10 @@
+package com.crypto.common.models
+
+@JvmInline
+value class CommonOrderId(private val id: String) {
+    fun asString() = id
+
+    companion object {
+        val NONE = CommonOrderId("")
+    }
+}
