+++ common/src/main/kotlin/com/crypto/common/models/CommonOrderOperation.kt
@@ -1,0 +1,7 @@
+package com.crypto.common.models
+
+enum class CommonOrderOperation {
+    NONE,
+    BUYING,
+    SELLING
+}
