+++ common/src/main/kotlin/com/crypto/common/models/CommonOrderStatus.kt
@@ -1,0 +1,7 @@
+package com.crypto.common.models
+
+enum class CommonOrderStatus {
+    NONE,
+    OPEN,
+    CLOSED
+}
