+++ common/src/main/kotlin/com/crypto/common/models/CommonOrderType.kt
@@ -1,0 +1,7 @@
+package com.crypto.common.models
+
+enum class CommonOrderType {
+    NONE,
+    MARKET,
+    LIMIT
+}
