+++ common/src/main/kotlin/com/crypto/common/models/CommonRequestId.kt
@@ -1,0 +1,10 @@
+package com.crypto.common.models
+
+@JvmInline
+value class CommonRequestId(private val id: String) {
+    fun asString() = id
+
+    companion object {
+        val NONE = CommonRequestId("")
+    }
+}
