+++ common/src/main/kotlin/com/crypto/common/models/CommonState.kt
@@ -1,0 +1,7 @@
+package com.crypto.common.models
+
+enum class CommonState {
+    NONE,
+    RUNNING,
+    FAILING
+}
