+++ common/src/main/kotlin/com/crypto/common/models/CommonWalletId.kt
@@ -1,0 +1,10 @@
+package com.crypto.common.models
+
+@JvmInline
+value class CommonWalletId(private val id: String) {
+    fun asString() = id
+
+    companion object {
+        val NONE = CommonWalletId("")
+    }
+}
