+++ common/src/main/kotlin/com/crypto/common/models/CommonWorkMode.kt
@@ -1,0 +1,7 @@
+package com.crypto.common.models
+
+enum class CommonWorkMode {
+    PROD,
+    TEST,
+    STUB,
+}
