+++ common/src/main/kotlin/com/crypto/common/stubs/CommonStub.kt
@@ -1,0 +1,12 @@
+package com.crypto.common.stubs
+
+enum class CommonStub {
+    NONE,
+    SUCCESS,
+    NOT_FOUND,
+    BAD_ID,
+    CANNOT_DELETE,
+    BAD_TYPE,
+    BAD_STATUS,
+    BAD_WALLET_ID
+}
