--- common/src/main/kotlin/models/CommonCommand.kt
@@ -1,9 +1,0 @@
-package models
-
-enum class CommonCommand {
-    NONE,
-    CREATE,
-    READ,
-    DELETE,
-    SEARCH
-}
