--- common/src/main/kotlin/models/CommonError.kt
@@ -1,9 +1,0 @@
-package models
-
-class CommonError(
-    val code: String = "",
-    val group: String = "",
-    val field: String = "",
-    val message: String = "",
-    val exception: Throwable? = null
-)
