--- common/src/main/kotlin/models/CommonFilterOrder.kt
@@ -1,7 +1,0 @@
-package models
-
-data class CommonFilterOrder(
-    var type: CommonOrderType = CommonOrderType.NONE,
-    var operation: CommonOrderOperation = CommonOrderOperation.NONE,
-    var status: CommonOrderStatus = CommonOrderStatus.NONE,
-)
