--- common/src/main/kotlin/models/CommonOrderId.kt
@@ -1,10 +1,0 @@
-package models
-
-@JvmInline
-value class CommonOrderId(private val id: String) {
-    fun asString() = id
-
-    companion object {
-        val NONE = CommonOrderId("")
-    }
-}
