--- common/src/main/kotlin/models/CommonOrderOperation.kt
@@ -1,7 +1,0 @@
-package models
-
-enum class CommonOrderOperation {
-    NONE,
-    BUYING,
-    SELLING
-}
