--- common/src/main/kotlin/models/CommonOrderStatus.kt
@@ -1,7 +1,0 @@
-package models
-
-enum class CommonOrderStatus {
-    NONE,
-    OPEN,
-    CLOSED
-}
