--- common/src/main/kotlin/models/CommonOrderType.kt
@@ -1,7 +1,0 @@
-package models
-
-enum class CommonOrderType {
-    NONE,
-    MARKET,
-    LIMIT
-}
