--- common/src/main/kotlin/models/CommonRequestId.kt
@@ -1,10 +1,0 @@
-package models
-
-@JvmInline
-value class CommonRequestId(private val id: String) {
-    fun asString() = id
-
-    companion object {
-        val NONE = CommonRequestId("")
-    }
-}
