--- common/src/main/kotlin/models/CommonState.kt
@@ -1,7 +1,0 @@
-package models
-
-enum class CommonState {
-    NONE,
-    RUNNING,
-    FAILING
-}
