--- common/src/main/kotlin/models/CommonWalletId.kt
@@ -1,10 +1,0 @@
-package models
-
-@JvmInline
-value class CommonWalletId(private val id: String) {
-    fun asString() = id
-
-    companion object {
-        val NONE = CommonWalletId("")
-    }
-}
