--- common/src/main/kotlin/models/CommonWorkMode.kt
@@ -1,7 +1,0 @@
-package models
-
-enum class CommonWorkMode {
-    PROD,
-    TEST,
-    STUB,
-}
