--- common/src/main/kotlin/stubs/CommonStub.kt
@@ -1,12 +1,0 @@
-package stubs
-
-enum class CommonStub {
-    NONE,
-    SUCCESS,
-    NOT_FOUND,
-    BAD_ID,
-    CANNOT_DELETE,
-    BAD_TYPE,
-    BAD_STATUS,
-    BAD_WALLET_ID
-}
