+++ gradle.properties
@@ -7,3 +7,8 @@ openapiVersion=6.3.0
 jacksonVersion=2.14.1
 
 kotestVersion=5.5.5
+
+springframeworkBootVersion=3.0.4
+springDependencyManagementVersion=1.1.0
+pluginSpringVersion=1.8.0
+springdocOpenapiUiVersion=2.0.4
