+++ mappers/src/main/kotlin/com/crypto/mappers/exceptions/UnknownCommandException.kt
@@ -1,0 +1,6 @@
+package com.crypto.mappers.exceptions
+
+import com.crypto.common.models.CommonCommand
+
+class UnknownCommandException(command: CommonCommand) : RuntimeException("Unknown command $command")
+
