+++ mappers/src/main/kotlin/com/crypto/mappers/exceptions/UnknownRequestException.kt
@@ -1,0 +1,5 @@
+package com.crypto.mappers.exceptions
+
+import kotlin.reflect.KClass
+
+class UnknownRequestException(kClass: KClass<*>) : RuntimeException("Class $kClass cannot be mapped and not supported")
