--- mappers/src/main/kotlin/exceptions/UnknownCommandException.kt
@@ -1,6 +1,0 @@
-package exceptions
-
-import models.CommonCommand
-
-class UnknownCommandException(command: CommonCommand) : RuntimeException("Unknown command $command")
-
