--- mappers/src/main/kotlin/exceptions/UnknownRequestException.kt
@@ -1,5 +1,0 @@
-package exceptions
-
-import kotlin.reflect.KClass
-
-class UnknownRequestException(kClass: KClass<*>) : RuntimeException("Class $kClass cannot be mapped and not supported")
