+++ order-app-spring/src/main/kotlin/com/crypto/orderappspring/OrderAppSpringApplication.kt
@@ -1,0 +1,11 @@
+package com.crypto.orderappspring
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class OrderAppSpringApplication
+
+fun main(args: Array<String>) {
+	runApplication<OrderAppSpringApplication>(*args)
+}
