+++ order-app-spring/src/main/kotlin/com/crypto/orderappspring/config/SwaggerConfiguration.kt
@@ -1,0 +1,20 @@
+package com.crypto.orderappspring.config
+
+import org.springdoc.core.configuration.SpringDocConfiguration
+import org.springdoc.core.properties.SpringDocConfigProperties
+import org.springdoc.core.providers.ObjectMapperProvider
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+class SwaggerConfiguration {
+    @Bean
+    fun springDocConfiguration() = SpringDocConfiguration()
+
+    @Bean
+    fun springDocConfigProperties() = SpringDocConfigProperties()
+
+    @Bean
+    fun objectMapperProvider(springDocConfigProperties: SpringDocConfigProperties) =
+        ObjectMapperProvider(springDocConfigProperties)
+}
