+++ order-app-spring/src/test/kotlin/com/crypto/orderappspring/OrderAppSpringApplicationTests.kt
@@ -1,0 +1,14 @@
+package com.crypto.orderappspring
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class OrderAppSpringApplicationTests {
+
+	@Test
+	fun contextLoads() {
+		println("Hi")
+	}
+
+}
