+++ settings.gradle.kts
@@ -2,14 +2,27 @@ rootProject.name = "crypto-exchange"
 
 pluginManagement {
     val kotlinVersion: String by settings
+    val kotestVersion: String by settings
     val openapiVersion: String by settings
+    val springframeworkBootVersion: String by settings
+    val springDependencyManagementVersion: String by settings
+    val pluginSpringVersion: String by settings
 
     plugins {
-        kotlin("jvm") version kotlinVersion
+        kotlin("jvm") version kotlinVersion apply false
+        id("io.kotest.multiplatform") version kotestVersion apply false
+        kotlin("plugin.serialization") version kotlinVersion apply false
+
         id("org.openapi.generator") version openapiVersion apply false
+
+        id("org.springframework.boot") version springframeworkBootVersion apply false
+        id("io.spring.dependency-management") version springDependencyManagementVersion apply false
+        kotlin("plugin.spring") version pluginSpringVersion apply false
     }
 }
 
 include("api")
 include("mappers")
 include("common")
+include("stubs")
+include("order-app-spring")
