+++ stubs/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    kotlin("jvm")
+}
+
+dependencies {
+    implementation(project(":common"))
+}
