+++ stubs/src/main/kotlin/com/crypto/stubs/OrderStub.kt
@@ -1,0 +1,7 @@
+package com.crypto.stubs
+
+import com.crypto.common.models.CommonOrder
+
+object OrderStub {
+    fun get(): CommonOrder = OrderStubUsdt.ORDER_BUYING_USDT.copy()
+}
